speculation-rules/prerender/prerender-response-codes.html

<!DOCTYPE html>
<title>Same-origin prerendering can access localStorage</title>
<meta name="timeout" content="long">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/utils.js"></script>
<script src="resources/utils.js"></script>
<body>
<script>

setup(() => assertSpeculationRulesIsSupported());
function test_prerender_response_code(code, allow) {
  promise_test(async t => {
    const uid = token();
    const url = `resources/prerender-response-code.html?code=${code}&uid=${uid}`;
    const bc = new BroadcastChannel(uid);
    t.add_cleanup(() => bc.close());
    const popup = window.open(url, '_blank', 'noopener');
    const result = await new Promise(resolve => {
      bc.addEventListener('message', e => {
        console.log(e)
        if (e.data === 'activated' || e.data === 'discarded')
          resolve(e.data);
      }, {once: true});
    });

    assert_equals(result, allow ? 'activated' : 'discarded');
  },`Responses with code ${code} should ${allow ? 'go ahead with' : 'discard'} prerendering`);
}

[200, 201, 202, 203].forEach(code => test_prerender_response_code(code, true));
[204, 205].forEach(code => test_prerender_response_code(code, false));
</script>
</body>
